+++ test_civic_jabber_ingest/external_services/test_aws.py
@@ -3,7 +3,7 @@ import subprocess
 
 import pytest
 
-import civic_jabber_ingest.utils.aws as aws
+import civic_jabber_ingest.external_services.aws as aws
 from civic_jabber_ingest.utils.environ import modified_environ
